Added light ticket purchase functionality
Ticket functionality works on the database level and can record purchases with the number of tickets, total cost, event ID and purchaser ID.

+++ a3_starter_code-main/projectfile/website/forms.py
@@ -58,7 +58,7 @@ class ValidEditForm(FlaskForm):
 # to be done
 # this form will be used to allow users to purchase tickets
 class TicketPurchase(FlaskForm):
-    numTickets = DecimalField('Number of tickets to purchase', validators=[InputRequired()])
+    numTickets = IntegerField('Number of tickets to purchase', validators=[InputRequired()])
     submit = SubmitField('Purchase Ticket(s)')
 
 # to be done
